src/app/shared/components/product/stock/stock.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { StockComponent } from './stock.component';

describe('Stock Component', () => {
  let component: StockComponent;
  let fixture: ComponentFixture<StockComponent>;
  let element: HTMLElement;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [StockComponent],
    }).compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(StockComponent);
    component = fixture.componentInstance;
    element = fixture.nativeElement;
  });

  it('should be created', () => {
    expect(component).toBeTruthy();
    expect(element).toBeTruthy();
    expect(() => fixture.detectChanges()).not.toThrow();
  });
});
